requirements.txt
fastapi==0.115.6
uvicorn[standard]==0.32.1
streamlit==1.38.0
requests==2.32.3
sentence-transformers==3.0.1
faiss-cpu==1.8.0
python-dotenv==1.0.1
pydantic==2.10.3
